# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FileRecordStatus = typeof FileRecordStatus[keyof typeof FileRecordStatus];


export const FileRecordStatus = {
  ready: 'ready',
  processing: 'processing',
  error: 'error',
} as const;

export interface FileRecord {
  id: string;
  name: string;
  kind: string;
  size: number;
  status: FileRecordStatus;
  createdAt: string;
  /** @nullable */
  pages?: number | null;
}

export interface FileInput {
  /** @minLength 1 */
  name: string;
  kind: string;
  /** @minimum 0 */
  size: number;
  /** @nullable */
  pages?: number | null;
  contentBase64?: string;
  mimeType?: string;
}

export interface Activity {
  id: string;
  label: string;
  detail: string;
  time: string;
}

export interface DashboardSummary {
  filesProcessed: number;
  storageUsed: string;
  aiRequests: number;
  plan: string;
  usagePercent: number;
  recentActivity: Activity[];
}

export interface ProcessingInput {
  fileId: string;
  operation: string;
  /** Optional user question for document-intelligence operations. */
  question?: string;
}

export type ProcessingJobStatus = typeof ProcessingJobStatus[keyof typeof ProcessingJobStatus];


export const ProcessingJobStatus = {
  queued: 'queued',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ProcessingJob {
  id: string;
  fileId: string;
  operation: string;
  status: ProcessingJobStatus;
  progress: number;
  message: string;
  resultUrl?: string;
  resultName?: string;
  resultContentType?: string;
}