# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';

export interface DashboardSummary {
  filesProcessed: number;
  storageUsed: string;
  aiRequests: number;
  plan: string;
  usagePercent: number;
  recentActivity: Activity[];
}