# lib/api-zod/src/generated/types/fileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileInput {
  /** @minLength 1 */
  name: string;
  kind: string;
  /** @minimum 0 */
  size: number;
  /** @nullable */
  pages?: number | null;
  contentBase64?: string;
  mimeType?: string;
}