# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileRecordStatus } from './fileRecordStatus';

export interface FileRecord {
  id: string;
  name: string;
  kind: string;
  size: number;
  status: FileRecordStatus;
  createdAt: string;
  /** @nullable */
  pages?: number | null;
}