# lib/api-zod/src/generated/types/processingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProcessingInput {
  fileId: string;
  operation: string;
  /** Optional user question for document-intelligence operations. */
  question?: string;
}