# lib/api-zod/src/generated/types/processingJob.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessingJobStatus } from './processingJobStatus';

export interface ProcessingJob {
  id: string;
  fileId: string;
  operation: string;
  status: ProcessingJobStatus;
  progress: number;
  message: string;
  resultUrl?: string;
  resultName?: string;
  resultContentType?: string;
}